(PA-5407) Simplify configure script
Remove most of the configure options in keeping with the OpenSSL 3 FIPS 140-2
Security Policy, Appendix A[1]. The options do affect which source files are
compiled, but not which algorithms are added to the fips module, which is the
only thing we install.

Most of the options will be applied to the `openssl-3.0` component in the
puppet-runtime project in PA-5409 (since that's where it matters).

Removed the `make depend` target, since it hasn't been necessary since OpenSSL 1.1.0[2]

[1] https://csrc.nist.gov/CSRC/media/projects/cryptographic-module-validation-program/documents/security-policies/140sp4282.pdf
[2] https://github.com/openssl/openssl/blob/openssl-3.0.0/INSTALL.md#makefile-targets