(PE-36616) Compile openssl-fips module for bolt-server
This commit adds a new project in the openssl-fips vanagon repo specifically to compile the module for use in bolt server. The important differences are that this only needs to be compiled for primary platforms (IE not windows) and use the correct install root and project prefix such that the library can be loaded correctly by bolt-server's ruby environment.